LayoutTests/http/tests/workers/service/basic-unregister.https-expected.txt

PASS: There is initially no service worker registered for the origin
PASS: registration scope is https://127.0.0.1:8443/
PASS: There is a service worker registered for the origin
PASS: Unregistration was successful
PASS: There is no service worker registered for the origin
PASS: Unregistration failed as expected
PASS: There is no service worker registered for the origin
PASS: registration scope is https://127.0.0.1:8443/workers/service/
PASS: There is a service worker registered for the origin
